Relocate abstract holders
