+++ api-bukkit/src/main/java/at/pavlov/bukkit/cannons/data/BukkitCannonDesign.java
@@ -1,10 +1,10 @@
 package at.pavlov.bukkit.cannons.data;
 
 import at.pavlov.bukkit.cannons.CannonBukkit;
-import at.pavlov.bukkit.container.BukkitBlock;
+import at.pavlov.bukkit.container.holders.BukkitBlock;
 import at.pavlov.bukkit.container.BukkitCannonBlocks;
-import at.pavlov.bukkit.container.BukkitItemHolder;
-import at.pavlov.bukkit.container.BukkitSoundHolder;
+import at.pavlov.bukkit.container.holders.BukkitItemHolder;
+import at.pavlov.bukkit.container.holders.BukkitSoundHolder;
 import at.pavlov.bukkit.factory.VectorUtils;
 import at.pavlov.internal.CannonLogger;
 import at.pavlov.internal.cannons.data.CannonDesign;
