+++ api-bukkit/src/main/java/at/pavlov/bukkit/projectile/BukkitProjectile.java
@@ -1,10 +1,10 @@
 package at.pavlov.bukkit.projectile;
 
 import at.pavlov.bukkit.builders.ParticleBuilder;
-import at.pavlov.bukkit.container.BukkitEntityHolder;
-import at.pavlov.bukkit.container.BukkitItemHolder;
-import at.pavlov.bukkit.container.BukkitSoundHolder;
-import at.pavlov.bukkit.container.BukkitSpawnMaterialHolder;
+import at.pavlov.bukkit.container.holders.BukkitEntityHolder;
+import at.pavlov.bukkit.container.holders.BukkitItemHolder;
+import at.pavlov.bukkit.container.holders.BukkitSoundHolder;
+import at.pavlov.bukkit.container.holders.BukkitSpawnMaterialHolder;
 import at.pavlov.internal.projectile.Projectile;
 import org.bukkit.FireworkEffect;
 import org.bukkit.block.data.BlockData;
