+++ api-internal/src/main/java/at/pavlov/internal/container/holders/ItemHolder.java
@@ -1,4 +1,4 @@
-package at.pavlov.internal.container;
+package at.pavlov.internal.container.holders;
 
 import lombok.Getter;
