+++ cannons-bukkit/src/main/java/at/pavlov/cannons/cannon/Cannon.java
@@ -2,9 +2,9 @@ package at.pavlov.cannons.cannon;
 
 import at.pavlov.bukkit.cannons.CannonBukkit;
 import at.pavlov.bukkit.cannons.data.BukkitCannonDesign;
-import at.pavlov.bukkit.container.BukkitBlock;
+import at.pavlov.bukkit.container.holders.BukkitBlock;
 import at.pavlov.bukkit.container.BukkitCannonBlocks;
-import at.pavlov.bukkit.container.BukkitItemHolder;
+import at.pavlov.bukkit.container.holders.BukkitItemHolder;
 import at.pavlov.bukkit.factory.CoordinateUtil;
 import at.pavlov.bukkit.projectile.BukkitProjectile;
 import at.pavlov.cannons.Cannons;
