+++ cannons-bukkit/src/main/java/at/pavlov/cannons/hooks/papi/CannonsPAPIExpansion.java
@@ -1,6 +1,6 @@
 package at.pavlov.cannons.hooks.papi;
 
-import at.pavlov.bukkit.container.BukkitItemHolder;
+import at.pavlov.bukkit.container.holders.BukkitItemHolder;
 import at.pavlov.cannons.Aiming;
 import at.pavlov.cannons.Cannons;
 import me.clip.placeholderapi.expansion.PlaceholderExpansion;
